+++ client/Sample/freerdp.c
@@ -139,7 +139,7 @@ static void* tf_client_thread_proc(freerdp* instance)
 		if (status == WAIT_FAILED)
 		{
 			WLog_ERR(TAG, "%s: WaitForMultipleObjects failed with %lu", __FUNCTION__,
-			         status);
+			         (unsigned long) status);
 			break;
 		}
