+++ client/X11/xf_window.c
@@ -238,7 +238,7 @@ BOOL xf_GetWindowProperty(xfContext* xfc, Window window, Atom property,
 
 	if (actual_type == None)
 	{
-		WLog_INFO(TAG, "Property %lu does not exist", property);
+		WLog_INFO(TAG, "Property %lu does not exist", (unsigned long) property);
 		return FALSE;
 	}
