+++ winpr/libwinpr/utils/debug.c
@@ -424,7 +424,7 @@ char** winpr_backtrace_symbols(void* buffer, size_t* used)
 
 			if (SymGetLineFromAddr64(process, address, &displacement, line))
 			{
-				sprintf_s(vlines[i], line_len, "%08lX: %s in %s:%lu", symbol->Address, symbol->Name, line->FileName, line->LineNumber);
+				sprintf_s(vlines[i], line_len, "%08lX: %s in %s:%lu", symbol->Address, symbol->Name, line->FileName, (unsigned long) line->LineNumber);
 			}
 			else
 				sprintf_s(vlines[i], line_len, "%08lX: %s", symbol->Address, symbol->Name);
